+++ www/Site/Library/admin_widgets/Hiding/Hiding.php
@@ -5,10 +5,18 @@
  */
 namespace Library\admin_widgets\Hiding;
 
+use Boolive\values\Rule;
 use Library\admin_widgets\ToggleAction\ToggleAction;
 
 class Hiding extends ToggleAction
 {
+    function startRule()
+    {
+        $rule = parent::startRule();
+        $rule->arrays[0]['REQUEST']->arrays[0]['select'] = Rule::in(null, 'structure', 'property', 'heirs', 'protos')->required();
+        return $rule;
+    }
+
     protected function initState()
     {
         /** @var \Boolive\data\Entity $object */