+++ www/Site/Library/admin_widgets/isMandatory/isMandatory.php
@@ -7,10 +7,18 @@
 namespace Library\admin_widgets\isMandatory;
 
 use Boolive\errors\Error;
+use Boolive\values\Rule;
 use Library\admin_widgets\ToggleAction\ToggleAction;
 
 class isMandatory extends ToggleAction
 {
+    function startRule()
+    {
+        $rule = parent::startRule();
+        $rule->arrays[0]['REQUEST']->arrays[0]['select'] = Rule::in(null, 'structure', 'property', 'heirs', 'protos')->required();
+        return $rule;
+    }
+
     protected function initState()
     {
         /** @var \Boolive\data\Entity $object */